Continued debugging, added "juice"
Added extra camera movement. Realised light is relative to camera, need to make it relative to models.

Struggling not having the noise texture over everything. Not fully understanding how to organise the buffers to achieve what I want. I need to save the noise texture onto the colour buffer so that it gets applied to the object and not as a filter, which is what is currently happening. Need to brush up on texture's tomorrow and just rrrrroll.