Cleaned Project. Added new model. Struggling to have 2 objects.
Downloaded a spaceship model off Turbosquid to add some intrigue. Not added to Git MUST REMEMBER TO ADD TO REPO MANUALLY.

Want to create shadow map as a final touch.

Struggling to render both plane and ship in same FBO. Need to do 2 different passes so that the ship isn't wobbly or projected onto the plane.